requirements_docs.txt
ase==3.22.1
gdown==4.6.0
matplotlib==3.6.2
numpydoc==1.5.0
pydata-sphinx-theme==0.12.0
pypandoc==1.10
pytest-sphinx==0.5.0
pyvista==0.37.0
scikit-image==0.19.3
scipy==1.10.0
seekpath==2.0.1
Sphinx==5.3.0
sphinx-copybutton==0.5.1
sphinx-gallery==0.11.1
sphinx-design==0.3.0
sympy==1.11.1
trimesh==3.18.0